+++ demo/chat/src/containers/messages/Messages.jsx
@@ -162,7 +162,13 @@ const Messages = ({
     let editedNote;
     if (m.LastEditedTimestamp && !m.Redacted) {
       const time = formatTime(m.LastEditedTimestamp);
-      const date = formatDate(m.LastEditedTimestamp);
+      const date = formatDate(
+        m.LastEditedTimestamp,
+        undefined,
+        undefined,
+        'today',
+        'yesterday'
+      );
       editedNote = (
         <i style={{ fontStyle: 'italic' }}>{` (edited ${date} at ${time})`}</i>
       );
@@ -257,21 +263,21 @@ const Messages = ({
             <ChatBubble
               variant={variant}
               senderName={m.senderName}
-              content={m.content}
               redacted={m.redacted}
               showName={showName}
               showTail={showTail}
             >
-              {m.editedNote}
-              {m.metadata ? (
+              <div>
+                {m.content}
+                {m.editedNote}
+              </div>
+              {m.metadata && (
                 <div style={{ marginTop: '10px' }}>
                   <AttachmentProcessor
                     senderId={m.senderId}
                     {...attachment(m.metadata)}
                   />
                 </div>
-              ) : (
-                ''
               )}
             </ChatBubble>
           )}