+++ src/providers/MeetingProvider/MeetingManager.ts
@@ -75,7 +75,7 @@ export class MeetingManager implements AudioVideoObserver {
 
   devicePermissionStatus = DevicePermissionStatus.UNSET;
 
-  devicePermissionsObservers: ((permission: string) => void)[] = [];
+  devicePermissionsObservers: ((permission: DevicePermissionStatus) => void)[] = [];
 
   activeSpeakerListener: ((activeSpeakers: string[]) => void) | null = null;
 
@@ -437,13 +437,13 @@ export class MeetingManager implements AudioVideoObserver {
   };
 
   subscribeToDevicePermissionStatus = (
-    callback: (permission: string) => void
+    callback: (permission: DevicePermissionStatus) => void
   ): void => {
     this.devicePermissionsObservers.push(callback);
   };
 
   unsubscribeFromDevicePermissionStatus = (
-    callbackToRemove: (permission: string) => void
+    callbackToRemove: (permission: DevicePermissionStatus) => void
   ): void => {
     this.devicePermissionsObservers = this.devicePermissionsObservers.filter(
       callback => callback !== callbackToRemove